Board builder
Alot of minor fixes here and there in this commit.
The major changes is that i have done a lot of work in the board class and tried to implement a builder to that. Also made it possible to print the board nicely in console by overriden the ToString method